requirements.txt
aioredis==1.3.1
amqp==2.6.0
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.2.5
defusedxml==0.7.0rc1
diff-match-patch==20181111
Django==2.2.12
django-allauth==0.42.0
django-anymail==7.1.0
django-appconf==1.0.4
django-celery-beat==2.0.0
django-compressor==2.4
django-contrib-comments==1.9.2
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.2.0
django-haystack==2.8.1
django-import-export==2.1.0
django-markdownx==3.0.1
django-reversion==3.0.7
django-taggit==1.2.0
django-taggit-serializer==0.1.7
django-timezone-field==4.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-extensions==0.6.0
drf-haystack==1.8.6
elasticsearch==1.9.0
et-xmlfile==1.0.1
future==0.18.2
hiredis==1.0.1
httplib2==0.9.2
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==4.6.10
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
msgpack==0.6.2
mysqlclient==1.4.6
nexmo==2.4.0
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
rcssmin==1.0.6
requests==2.23.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
service-identity==18.1.0
simplejson==3.17.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
tablib==1.1.0
text-unidecode==1.3
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
uWSGI==2.0.18
vine==1.3.0
Werkzeug==1.0.1
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.1.0
